test: add test cases for manual mode
