+++ anthos-bm-gcp-terraform/test/unit/main_custom_mode_test.go
@@ -22,11 +22,97 @@ import (
 	"testing"
 
 	"github.com/GoogleCloudPlatform/anthos-samples/anthos-bm-gcp-terraform/util"
+	"github.com/GoogleCloudPlatform/anthos-samples/anthos-bm-gcp-terraform/validation"
 	"github.com/gruntwork-io/terratest/modules/gcp"
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	testStructure "github.com/gruntwork-io/terratest/modules/test-structure"
 	"github.com/stretchr/testify/assert"
 )
+
+func TestUnit_MainScript_InstallMode(goTester *testing.T) {
+	moduleDir := testStructure.CopyTerraformFolderToTemp(goTester, "../../", ".")
+	projectID := gcp.GetGoogleProjectIDFromEnvVar(goTester) // from GOOGLE_CLOUD_PROJECT
+
+	workingDir, err := os.Getwd()
+	util.LogError(err, "Failed to read current working directory")
+	credentialsFile := fmt.Sprintf("%s/credentials_file.json", workingDir)
+
+	tmpFile, err := os.Create(credentialsFile)
+	util.LogError(err, fmt.Sprintf("Could not create temporary file at %s", credentialsFile))
+	defer tmpFile.Close()
+	defer os.Remove(credentialsFile)
+
+	mode := "install"
+	resourcesPath := "./resources"
+	bootDiskSize := 175
+	instanceCount := map[string]int{
+		"controlplane": 3,
+		"worker":       2,
+	}
+
+	tfPlanOutput := "terraform_test.tfplan"
+	tfPlanOutputArg := fmt.Sprintf("-out=%s", tfPlanOutput)
+	tfVarsMap := map[string]interface{}{
+		"project_id":       projectID,
+		"credentials_file": credentialsFile,
+		"mode":             mode,
+		"resources_path":   resourcesPath,
+		"instance_count":   instanceCount,
+		"boot_disk_size":   bootDiskSize,
+	}
+
+	tfOptions := terraform.WithDefaultRetryableErrors(goTester, &terraform.Options{
+		TerraformDir: moduleDir,
+		// Variables to pass to our Terraform code using -var options
+		Vars:         tfVarsMap,
+		PlanFilePath: tfPlanOutput,
+	})
+
+	// Terraform init and plan only
+	terraform.Init(goTester, tfOptions)
+	terraform.RunTerraformCommand(
+		goTester,
+		tfOptions,
+		terraform.FormatArgs(tfOptions, "plan", tfPlanOutputArg)...,
+	)
+	tfPlanJSON, err := terraform.ShowE(goTester, tfOptions)
+	util.LogError(err, fmt.Sprintf("Failed to parse the plan file %s into JSON format", tfPlanOutput))
+
+	var terraformPlan util.MainModulePlan
+	err = json.Unmarshal([]byte(tfPlanJSON), &terraformPlan)
+	util.LogError(err, "Failed to parse the JSON plan into the MainModulePlan struct in unit/module_main.go")
+	/**
+	 * Pro tip:
+	 * Write the json to a file using the util.WriteToFile() method to easily debug
+	 * util.WriteToFile(tfPlanJSON, "../../plan.json")
+	 */
+
+	var installAbmModule []util.TFModule // install_abm
+
+	for _, childModule := range terraformPlan.PlannedValues.RootModule.ChildModules {
+		moduleAddress := childModule.ModuleAddress
+		if strings.HasSuffix(moduleAddress, "instance_template") ||
+			strings.HasSuffix(moduleAddress, "vm_hosts") ||
+			strings.HasSuffix(moduleAddress, "service_accounts") ||
+			strings.Contains(moduleAddress, "google_apis") ||
+			strings.Contains(moduleAddress, "init_hosts") {
+			continue
+		} else if strings.Contains(moduleAddress, "install_abm") {
+			installAbmModule = append(installAbmModule, childModule)
+		} else {
+			goTester.Errorf("Unexpected module with address [%s] at planned_values.root_module.child_modules", moduleAddress)
+		}
+	}
+
+	assert.Len(
+		goTester,
+		installAbmModule,
+		1,
+		"Unexpected number of child modules with address type install_abm at planned_values.root_module.child_modules",
+	)
+	// validate the outputs from the script
+	validation.ValidateMainOutputsForInstallMode(goTester, terraformPlan.PlannedValues.Outputs, &tfVarsMap)
+}
 
 func TestUnit_MainScript_ManualLB(goTester *testing.T) {
 	moduleDir := testStructure.CopyTerraformFolderToTemp(goTester, "../../", ".")
@@ -117,13 +203,24 @@ func TestUnit_MainScript_ManualLB(goTester *testing.T) {
 
 	var ingressLBModule []util.TFModule
 	var controlplaneLBModule []util.TFModule
+	var installAbmModule []util.TFModule // install_abm
 
 	for _, childModule := range terraformPlan.PlannedValues.RootModule.ChildModules {
 		moduleAddress := childModule.ModuleAddress
-		if strings.HasSuffix(moduleAddress, "configure_ingress_lb[0]") {
+		if strings.HasSuffix(moduleAddress, "instance_template") ||
+			strings.HasSuffix(moduleAddress, "vm_hosts") ||
+			strings.HasSuffix(moduleAddress, "service_accounts") ||
+			strings.Contains(moduleAddress, "google_apis") ||
+			strings.Contains(moduleAddress, "init_hosts") {
+			continue
+		} else if strings.HasSuffix(moduleAddress, "configure_ingress_lb[0]") {
 			ingressLBModule = append(ingressLBModule, childModule)
 		} else if strings.HasSuffix(moduleAddress, "configure_controlplane_lb[0]") {
 			controlplaneLBModule = append(controlplaneLBModule, childModule)
+		} else if strings.Contains(moduleAddress, "install_abm") {
+			installAbmModule = append(installAbmModule, childModule)
+		} else {
+			goTester.Errorf("Unexpected module with address [%s] at planned_values.root_module.child_modules", moduleAddress)
 		}
 	}
 
@@ -139,4 +236,13 @@ func TestUnit_MainScript_ManualLB(goTester *testing.T) {
 		1,
 		"Unexpected number of child modules with address type configure_controlplane_lb at planned_values.root_module.child_modules",
 	)
+	assert.Len(
+		goTester,
+		installAbmModule,
+		1,
+		"Unexpected number of child modules with address type install_abm at planned_values.root_module.child_modules",
+	)
+
+	// validate the outputs from the script
+	validation.ValidateMainOutputsForManualMode(goTester, terraformPlan.PlannedValues.Outputs, &tfVarsMap)
 }
