+++ anthos-bm-gcp-terraform/test/util/module_main.go
@@ -61,8 +61,10 @@ type MainPlannedValues struct {
 
 // Outputs represents the outputs produced by the main terraform module
 type Outputs struct {
-	AdminVMSSH   *AdminVMSSH   `json:"admin_vm_ssh"`
-	InstallCheck *InstallCheck `json:"installation_check"`
+	AdminVMSSH     *AdminVMSSH   `json:"admin_vm_ssh"`
+	InstallCheck   *InstallCheck `json:"installation_check"`
+	ControlPlaneIP *Variable     `json:"controlplane_ip"`
+	IngressIP      *Variable     `json:"ingress_ip"`
 }
 
 // AdminVMSSH represents the final output from the main terraform script with
