+++ anthos-bm-gcp-terraform/test/validation/main_test_validation.go
@@ -905,6 +905,69 @@ func ValidateMainOutputsForInstallMode(goTester *testing.T, planOutputs *util.Ou
 		"Module is not expected to have an output for admin_vm_ssh; but found",
 	)
 
+	// verify module does NOT produce an output for controlplane_ip in plan
+	assert.Equal(
+		goTester,
+		planOutputs.ControlPlaneIP.Value,
+		"",
+		"Module is not expected to have an output for controlplane_ip; but found",
+	)
+
+	// verify module does NOT produce an output for ingress_ip in plan
+	assert.Equal(
+		goTester,
+		planOutputs.IngressIP.Value,
+		"",
+		"Module is not expected to have an output for ingress_ip; but found",
+	)
+
+	outputValue := planOutputs.InstallCheck.Value
+	assert.True(
+		goTester,
+		strings.Contains(outputValue, "tail -f ~/install_abm.log"),
+		"Output is expected to have 'tail -f ~/install_abm.log'",
+	)
+}
+
+// ValidateMainOutputsForManualMode validates if the outputs in the terraform plan matches
+// the outputs defined in the `output.tf` of the main module for the mode "manuallb".
+func ValidateMainOutputsForManualMode(goTester *testing.T, planOutputs *util.Outputs, vars *map[string]interface{}) {
+	// verify module produces output in plan
+	assert.NotNil(
+		goTester,
+		planOutputs,
+		"Module is expected to produce outputs; but none found",
+	)
+
+	// verify module produces an output for installation_check in plan
+	assert.NotNil(
+		goTester,
+		planOutputs.InstallCheck,
+		"Module is expected to have an output for installation_check; but not found",
+	)
+
+	// verify module produces an output for controlplane_ip in plan
+	assert.NotNil(
+		goTester,
+		planOutputs.ControlPlaneIP,
+		"Module is expected to have an output for controlplane_ip; but not found",
+	)
+
+	// verify module produces an output for ingress_ip in plan
+	assert.NotNil(
+		goTester,
+		planOutputs.IngressIP,
+		"Module is expected to have an output for ingress_ip; but not found",
+	)
+
+	// verify module does NOT produce an output for admin_vm_ssh in plan
+	assert.Equal(
+		goTester,
+		"",
+		planOutputs.AdminVMSSH.Value,
+		"Module is not expected to have an output for admin_vm_ssh; but found",
+	)
+
 	outputValue := planOutputs.InstallCheck.Value
 	assert.True(
 		goTester,
